Made GetMCS crd a common method.
